🐛 Bugfix: [Add explicit float64 types to numpy array creations] (#355)
💡 What
Added explicit `dtype=np.float64` parameters to `np.array()`
initializations across several backend files (`goal_planner_agent.py`,
`quant_engine.py`, `portfolio_analyzer.py`, `market_routes.py`). Marked
suggestion 020 as `Completed` in `future/suggestions.MD`.

🎯 Why
When creating NumPy arrays from dynamic input data, omitting the `dtype`
allows NumPy to implicitly infer the type. If the input happens to
contain only integers (e.g., mock data, perfect whole numbers in prices
or residuals), the array becomes `int64`. Subsequent operations that
require floats (like division or bounded constrained optimization loops)
will crash with a `_UFuncOutputCastingError`. Explicitly enforcing
`np.float64` guarantees type safety for downstream mathematical
processing.

📸 Before/After
Before: `init_guess = np.array([1.0/n] * n)`
After: `init_guess = np.array([1.0/n] * n, dtype=np.float64)`

♿ Accessibility
N/A - Backend logic change.

---
*PR created automatically by Jules for task
[195792293594153971](https://jules.google.com/task/195792293594153971)
started by @sank29mane*

+++ backend/routes/market_routes.py
@@ -313,7 +313,7 @@ async def get_portfolio_history(days: int = 30, account_type: Optional[str] = No
 
                     # 3. Calculate portfolio value using dot product
                     ordered_tickers = list(valid_holdings.keys())
-                    weights = np.array([valid_holdings[t] for t in ordered_tickers])
+                    weights = np.array([valid_holdings[t] for t in ordered_tickers], dtype=np.float64)
 
                     # Subset dataframe
                     df_subset = df_calc[ordered_tickers]
